RENAME, DESIGN : 네비게이션 바 재구성 -> id수정
네비게이션 바 재구성에 따라
- image 변경
- fragment명 수정
- action명 수정
