--- app/src/main/java/com/example/blueleaf/fragments/StoreFragment.kt
@@ -1,48 +1,0 @@
-package com.example.blueleaf.fragments
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.databinding.DataBindingUtil
-import androidx.navigation.findNavController
-import com.example.blueleaf.R
-import com.example.blueleaf.databinding.FragmentStoreBinding
-
-
-class StoreFragment : Fragment() {
-    private lateinit var binding: FragmentStoreBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        binding = DataBindingUtil.inflate(inflater,R.layout.fragment_store,container,false)
-
-        binding.homeTab.setOnClickListener(){
-            it.findNavController().navigate(R.id.action_storeFragment_to_homeFragment)
-        }
-
-        binding.plantTab.setOnClickListener(){
-            it.findNavController().navigate(R.id.action_storeFragment_to_plantFragment)
-        }
-
-        binding.boardTab.setOnClickListener(){
-            it.findNavController().navigate(R.id.action_storeFragment_to_boardFragment)
-        }
-
-        binding.bookmarkTab.setOnClickListener(){
-            it.findNavController().navigate(R.id.action_storeFragment_to_bookmarkFragment)
-        }
-
-        return binding.root
-    }
-
-
-}
