+++ app/src/main/res/layout/fragment_store.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".fragments.StoreFragment">
+        tools:context=".fragments.PlantFragment">
 
         <!-- TODO: Update blank fragment layout -->
         <TextView
@@ -73,11 +73,11 @@
                     android:src="@drawable/bottom_home" />
 
                 <ImageView
-                    android:id="@+id/plantTab"
+                    android:id="@+id/informationTab"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_weight="1"
-                    android:src="@drawable/bottom_plant" />
+                    android:src="@drawable/bottom_information" />
 
                 <ImageView
                     android:id="@+id/boardTab"
@@ -94,11 +94,11 @@
                     android:src="@drawable/bottom_bookmark" />
 
                 <ImageView
-                    android:id="@+id/storeTab"
+                    android:id="@+id/plantTab"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_weight="1"
-                    android:src="@drawable/bottom_store" />
+                    android:src="@drawable/bottom_plant" />
 
 
             </LinearLayout>
